agent-marketer/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
requests
pyjwt
